pyproject.toml
[project]
name = "sber-test-task"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.16.5",
    "asyncpg>=0.30.0",
    "bcrypt==4.0.1",
    "fastapi[all]>=0.116.1",
    "gunicorn>=23.0.0",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.10.1",
    "pyjwt>=2.10.1",
    "sqlalchemy>=2.0.43",
    "uvicorn>=0.35.0",
]

[project.optional-dependencies]
dev = [
    "ruff", 
    "pytest", 
    "pytest-asyncio", 
    "httpx", 
    "testcontainers[postgres]"
]


[tool.ruff]
line-length = 88
target-version = "py310"  
per-file-ignores = { "__init__.py" = ["F401"] }

# Исключения
exclude = [
    "__init__",
    ".bzr",
    ".direnv", 
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "A"]
ignore = ["B008"]

[tool.ruff.format]
quote-style = "double"


[dependency-groups]
dev = [
    "pytest-asyncio>=1.1.0",
    "pytest-mock>=3.15.0",
    "pytest-order>=1.3.0",
    "ruff>=0.12.8",
    "testcontainers[postgres]>=4.13.0",
]
